requirements.txt
discord-py
